requirements.txt
streamlit==1.31.1
requests==2.31.0
pandas==2.2.0
Pillow==11.2.1
python-whois==0.8.0
dnspython==2.6.1
beautifulsoup4==4.12.3
holehe==1.61
python-magic==0.4.27
python-magic-bin==0.4.14; platform_system == "Windows"
folium==0.15.1
geoip2==4.8.0
aiohttp==3.9.3
tqdm==4.66.6
colorama==0.4.6
maxminddb==2.5.2
python-nmap==0.7.1
tweepy==4.14.0
exifread==3.0.0
pycountry==23.12.11
h8mail==2.5.6
sherlock-project==0.1.0
social-analyzer==0.45
spiderfoot==4.0.0
whatweb==0.5.5
sublist3r==1.0
photon==1.3.0
exiftool==0.5.3
streamlit-option-menu==0.3.12
streamlit-extras==0.3.6
filetype==1.2.0
setuptools>=69.0.3
bs4>=0.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3 